Comment Update
+ Updated credit comments for my code
