+++ Assets/Scripts/Loot Table/LootInterface.cs
@@ -2,6 +2,7 @@ using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
+// William Min
 public interface LootInterface
 {
     public List<string> GenerateLootDrops();
