+++ Assets/Scripts/Loot Table/LootTable.cs
@@ -1,13 +1,13 @@
 using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
-using UnityEditor;
 
+// William Min
 [CreateAssetMenu]
 public class LootTable : ScriptableObject, LootInterface
 {
     [SerializeField]
-    public List<LootDropScript> Items;
+    private List<LootDropScript> Items;
 
     public List<string> GenerateLootDrops()
     {
@@ -44,23 +44,5 @@ public class LootTable : ScriptableObject, LootInterface
         }
 
         Debug.Log(result);
-    }
-}
-
-[CustomEditor(typeof(LootTable))]
-[CanEditMultipleObjects]
-public class LootTableEditor : Editor
-{
-    public override void OnInspectorGUI()
-    {
-        DrawDefaultInspector();
-
-        var tool = target as LootTable;
-        serializedObject.Update();
-
-        if (GUILayout.Button("Generate Loot"))
-        {
-            tool.DisplayLootDrops();
-        }
     }
 }
