+++ Assets/Scripts/Loot Table/LootTableEditor.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 69a0b90e4bf0e044a9838ba225cc5264
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
